refactor(sandbox): boot via a small dedicated helper bin, not the fat binary
The boot child used to be a re-invoked copy of the consumer's own binary (link-
time ctor + boot_if_requested), so `Sandbox::exec` copied and ad-hoc-signed a
40-115MB executable per build and re-signed it on every consumer rebuild.

Move the whole host-side VM boot path into a new `crates/ailoy-krun-boot` crate
(msb_krun + cortex virtio-fs serving + smoltcp net only, no LLM/agent/http/image
stack), built release+LTO by build.rs and embedded. `boot_helper_exe` now
materializes + signs that embedded ~5MB helper instead of the consumer binary,
and `exec` invokes it directly. Drops the ctor and the re-invoke model entirely.

Consequences:
- Signed boot helper: ~5MB (was 40-115MB); identical regardless of consumer, so
  no more per-consumer-rebuild churn/accumulation in ~/.ailoy/krun/boot-helper.
- ailoy the library no longer links msb_krun / microsandbox-network / rustls /
  ctor or cortex's krun binding — they live only in the boot helper. (ailoy's
  Cargo.lock shrinks accordingly; agent-k's will too.)
- boot_if_requested and the ctor are removed. The env/JSON wire contract between
  parent and the boot bin mirrors the existing guest-init split.

build.rs now compiles two helpers (guest-init musl + krun-boot host release);
the krun-boot build links cortex, so a clean ailoy build is slower (cached
incrementally). Verified: two sandboxes boot alpine concurrently, per-sandbox
init-root isolation + cleanup, distinct signed helpers.
